gridfs avatars without tmp files
